refactor: unify connection pool implementation
