+++ database/intelligent_connection_pool.py
@@ -1,22 +1,49 @@
 from __future__ import annotations
 
-"""Adaptive connection pool that adjusts size based on usage.
-
-The :class:`IntelligentConnectionPool` automatically expands or shrinks
-its pool of connections depending on current demand. Use ``acquire`` and
-``release`` like a standard connection pool to obtain connections while
-metrics are tracked for analysis.
-"""
+"""Thread-safe connection pool with adaptive sizing and circuit breaking."""
 
 import threading
 import time
-from typing import Callable, Dict, List, Tuple, Any
+from typing import Any, Callable, Dict, List, Tuple
 
+from config.database_exceptions import ConnectionValidationFailed
 from database.types import DatabaseConnection
 
 
+class CircuitBreaker:
+    """Simple circuit breaker implementation."""
+
+    def __init__(self, failure_threshold: int = 5, recovery_timeout: int = 30) -> None:
+        self.failure_threshold = failure_threshold
+        self.recovery_timeout = recovery_timeout
+        self._failures = 0
+        self._opened_at: float | None = None
+        self._lock = threading.Lock()
+
+    def record_success(self) -> None:
+        with self._lock:
+            self._failures = 0
+            self._opened_at = None
+
+    def record_failure(self) -> None:
+        with self._lock:
+            self._failures += 1
+            if self._failures >= self.failure_threshold:
+                self._opened_at = time.time()
+
+    def allows_request(self) -> bool:
+        with self._lock:
+            if self._opened_at is None:
+                return True
+            if time.time() - self._opened_at >= self.recovery_timeout:
+                self._failures = 0
+                self._opened_at = None
+                return True
+            return False
+
+
 class IntelligentConnectionPool:
-    """Connection pool with dynamic scaling and simple metrics."""
+    """Connection pool with dynamic scaling, metrics, and circuit breaking."""
 
     def __init__(
         self,
@@ -26,8 +53,12 @@ class IntelligentConnectionPool:
         timeout: int,
         shrink_timeout: int,
         *,
+        failure_threshold: int = 5,
+        recovery_timeout: int = 30,
         threshold: float = 0.75,
     ) -> None:
+        self._lock = threading.RLock()
+        self._condition = threading.Condition(self._lock)
         self._factory = factory
         self._min_size = min_size
         self._max_pool_size = max(max_size, min_size)
@@ -38,10 +69,12 @@ class IntelligentConnectionPool:
 
         self._pool: List[Tuple[DatabaseConnection, float]] = []
         self._active = 0
-        self._lock = threading.RLock()
-        self._metrics: Dict[str, Any] = {
+        self.circuit_breaker = CircuitBreaker(failure_threshold, recovery_timeout)
+        self.metrics: Dict[str, Any] = {
             "acquired": 0,
             "released": 0,
+            "failed": 0,
+            "timeouts": 0,
             "expansions": 0,
             "shrinks": 0,
             "acquire_times": [],
@@ -54,24 +87,23 @@ class IntelligentConnectionPool:
 
     # ------------------------------------------------------------------
     def _maybe_expand(self) -> None:
-        usage = (self._active - len(self._pool)) / max(self._max_size, 1)
+        if self._max_size == 0:
+            return
+        usage = (self._active - len(self._pool)) / self._max_size
         if usage >= self._threshold and self._max_size < self._max_pool_size:
             self._max_size = min(self._max_size * 2, self._max_pool_size)
-            self._metrics["expansions"] += 1
+            self.metrics["expansions"] += 1
 
     # ------------------------------------------------------------------
     def _shrink_idle_connections(self) -> None:
         now = time.time()
         new_pool: List[Tuple[DatabaseConnection, float]] = []
         for conn, ts in self._pool:
-            if (
-                now - ts > self._shrink_timeout
-                and self._max_size > self._min_size
-            ):
+            if now - ts > self._shrink_timeout and self._max_size > self._min_size:
                 conn.close()
                 self._active -= 1
                 self._max_size -= 1
-                self._metrics["shrinks"] += 1
+                self.metrics["shrinks"] += 1
             else:
                 new_pool.append((conn, ts))
         self._pool = new_pool
@@ -79,9 +111,11 @@ class IntelligentConnectionPool:
     # ------------------------------------------------------------------
     def get_connection(self) -> DatabaseConnection:
         start = time.time()
+        if not self.circuit_breaker.allows_request():
+            raise ConnectionValidationFailed("circuit open")
         deadline = start + self._timeout
-        while True:
-            with self._lock:
+        with self._condition:
+            while True:
                 self._shrink_idle_connections()
                 self._maybe_expand()
 
@@ -90,29 +124,49 @@ class IntelligentConnectionPool:
                     if not conn.health_check():
                         conn.close()
                         self._active -= 1
+                        self.metrics["failed"] += 1
+                        self.circuit_breaker.record_failure()
+                        if not self.circuit_breaker.allows_request():
+                            raise ConnectionValidationFailed("circuit open")
                         continue
-                    self._metrics["acquired"] += 1
-                    self._metrics["acquire_times"].append(time.time() - start)
+                    self.metrics["acquired"] += 1
+                    self.metrics["acquire_times"].append(time.time() - start)
+                    self.circuit_breaker.record_success()
+                    self._condition.notify()
                     return conn
 
                 if self._active < self._max_size:
                     conn = self._factory()
+                    if not conn.health_check():
+                        conn.close()
+                        self.metrics["failed"] += 1
+                        self.circuit_breaker.record_failure()
+                        if not self.circuit_breaker.allows_request():
+                            raise ConnectionValidationFailed("circuit open")
+                        continue
                     self._active += 1
-                    self._metrics["acquired"] += 1
-                    self._metrics["acquire_times"].append(time.time() - start)
+                    self.metrics["acquired"] += 1
+                    self.metrics["acquire_times"].append(time.time() - start)
+                    self.circuit_breaker.record_success()
+                    self._condition.notify()
                     return conn
 
-            if time.time() >= deadline:
-                raise TimeoutError("No available connection in pool")
-            time.sleep(0.05)
+                remaining = deadline - time.time()
+                if remaining <= 0:
+                    self.metrics["timeouts"] += 1
+                    self.circuit_breaker.record_failure()
+                    raise TimeoutError("No available connection in pool")
+                self._condition.wait(timeout=remaining)
 
     # ------------------------------------------------------------------
     def release_connection(self, conn: DatabaseConnection) -> None:
-        with self._lock:
+        with self._condition:
             self._shrink_idle_connections()
             if not conn.health_check():
                 conn.close()
                 self._active -= 1
+                self.metrics["failed"] += 1
+                self.circuit_breaker.record_failure()
                 return
 
             if len(self._pool) >= self._max_size:
@@ -120,11 +174,36 @@ class IntelligentConnectionPool:
                 self._active -= 1
             else:
                 self._pool.append((conn, time.time()))
-            self._metrics["released"] += 1
+            self.metrics["released"] += 1
+            self._condition.notify()
+
+    # ------------------------------------------------------------------
+    def health_check(self) -> bool:
+        with self._lock:
+            temp: List[Tuple[DatabaseConnection, float]] = []
+            healthy = True
+            while self._pool:
+                conn, ts = self._pool.pop()
+                if not conn.health_check():
+                    healthy = False
+                    conn.close()
+                    self._active -= 1
+                    self.metrics["failed"] += 1
+                    if self._max_size > self._min_size:
+                        self._max_size -= 1
+                else:
+                    temp.append((conn, ts))
+            for item in temp:
+                self.release_connection(item[0])
+            if not healthy:
+                self.circuit_breaker.record_failure()
+            else:
+                self.circuit_breaker.record_success()
+            return healthy
 
     # ------------------------------------------------------------------
     def get_metrics(self) -> Dict[str, Any]:
-        data = self._metrics.copy()
+        data = self.metrics.copy()
         times = data.pop("acquire_times")
         data["avg_acquire_time"] = (sum(times) / len(times)) if times else 0.0
         data["active"] = self._active
@@ -132,4 +211,4 @@ class IntelligentConnectionPool:
         return data
 
 
-__all__ = ["IntelligentConnectionPool"]
+__all__ = ["IntelligentConnectionPool", "CircuitBreaker"]
